% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{as.tags.reactable}
\alias{as.tags.reactable}
\title{Convert a reactable widget to HTML tags}
\usage{
\method{as.tags}{reactable}(x, standalone = FALSE)
}
\arguments{
\item{x}{a \code{\link[=reactable]{reactable()}} instance.}

\item{standalone}{}
}
\description{
This S3 method exists to enable \code{\link[=reactable]{reactable()}}'s \code{static} rendering option.
}
