% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-df.R
\name{backendDf}
\alias{backendDf}
\title{Data Frame Server Backend}
\usage{
backendDf()
}
\value{
A backend object to pass to the \code{backend} argument of \code{\link[=reactable]{reactable()}}.
}
\description{
Uses pure R data frame operations for table data processing in Shiny apps.
Does not require the V8 package.
}
\examples{
\dontrun{
reactable(data, backend = backendDf())
}

}
