% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-dt.R
\name{backendDt}
\alias{backendDt}
\title{data.table Server Backend}
\usage{
backendDt()
}
\value{
A backend object to pass to the \code{backend} argument of \code{\link[=reactable]{reactable()}}.
}
\description{
Uses the data.table package for table data processing in Shiny apps.
Requires the data.table package.
}
\examples{
\dontrun{
reactable(data, backend = backendDt())
}

}
