% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-v8.R
\name{backendV8}
\alias{backendV8}
\title{V8 Server Backend}
\usage{
backendV8()
}
\value{
A backend object to pass to the \code{backend} argument of \code{\link[=reactable]{reactable()}}.
}
\description{
Uses V8 (server-side JavaScript) for table data processing in Shiny apps.
This is the default server-side backend.
}
\examples{
\dontrun{
reactable(data, backend = backendV8())
}

}
