% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{colDef}
\alias{colDef}
\title{Column definitions}
\usage{
colDef(name = NULL, aggregate = NULL, sortable = NULL,
  resizable = NULL, filterable = NULL, show = TRUE,
  defaultSortOrder = NULL, format = NULL, render = NULL,
  minWidth = NULL, maxWidth = NULL, width = NULL, align = NULL,
  class = NULL, style = NULL, headerClass = NULL,
  headerStyle = NULL)
}
\arguments{
\item{name}{Column name.}

\item{aggregate}{Aggregate function name or \code{JS()} function.}

\item{sortable}{Enable sorting? Overrides the table option.}

\item{resizable}{Enable column resizing? Overrides the table option.}

\item{filterable}{Enable column filtering? Overrides the table option.}

\item{show}{Show the column? Defaults to \code{TRUE}.}

\item{defaultSortOrder}{Default sort order. Either \code{"asc"} for ascending
order or \code{"desc"} for descending order. Overrides the table option.}

\item{format}{Column formatting options. A \code{colFormat()} object to
format all cells, or a named list of \code{colFormat()} objects to format standard
cells (\code{"cell"}) and aggregated cells (\code{"aggregated"}) separately.}

\item{render}{Custom column renderer. A \code{JS()} function to render all cells,
or a named list of \code{JS()} functions to render standard cells (\code{"cell"}) and
aggregated cells (\code{"aggregated"}) separately.}

\item{minWidth}{Min width of the column in pixels.}

\item{maxWidth}{Max width of the column in pixels.}

\item{width}{Fixed width of the column in pixels. Overrides minWidth and maxWidth.}

\item{align}{Column alignment. One of \code{"left"}, \code{"right"}, \code{"center"}.}

\item{class}{Additional CSS classes to apply to cells.}

\item{style}{Named list of inline styles to apply to cells.}

\item{headerClass}{Additional CSS classes to apply to the header.}

\item{headerStyle}{Named list of inline styles to apply to the header.}
}
\description{
Column definitions
}
