% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{colFormat}
\alias{colFormat}
\title{Column formatting options}
\usage{
colFormat(prefix = NULL, suffix = NULL, digits = NULL,
  separators = FALSE, percent = FALSE, currency = NULL,
  locales = NULL)
}
\arguments{
\item{prefix}{Prefix string.}

\item{suffix}{Suffix string.}

\item{digits}{Max number of decimal places to round numbers.}

\item{separators}{Whether to use grouping separators for numbers, such as
thousands separators or thousand/lakh/crore separators. The format is
locale-dependent. Defaults to \code{FALSE}.}

\item{percent}{Format as a percentage?}

\item{currency}{Currency format. An ISO 4217 currency code such as \code{"USD"}
for the US dollar, \code{"EUR"} for the euro, or \code{"CNY"} for the Chinese RMB.
The format is locale-dependent.}

\item{locales}{Locales to use for number formatting. A character vector of
BCP 47 language tags, such as \code{"en-US"} for English (United States),
\code{"hi"} for Hindi, or \code{"sv-SE"} for Swedish (Sweden). Defaults to the locale
of the browser.}
}
\description{
Column formatting options
}
