% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{colGroup}
\alias{colGroup}
\title{Column group definitions}
\usage{
colGroup(name = NULL, columns = NULL, align = NULL,
  headerClass = NULL, headerStyle = NULL)
}
\arguments{
\item{name}{Column group name.}

\item{columns}{Character vector of column names in the group.}

\item{align}{Column group header alignment. One of \code{"left"}, \code{"right"}, \code{"center"}.}

\item{headerClass}{Additional CSS classes to apply to the header.}

\item{headerStyle}{Inline styles to apply to the header. A named list or
character string.}
}
\description{
Column group definitions
}
\examples{
reactable(
  iris,
  columns = list(
    Sepal.Length = colDef(name = "Length"),
    Sepal.Width = colDef(name = "Width"),
    Petal.Length = colDef(name = "Length"),
    Petal.Width = colDef(name = "Width")
  ),
  columnGroups = list(
    colGroup(name = "Sepal", columns = c("Sepal.Length", "Sepal.Width")),
    colGroup(name = "Petal", columns = c("Petal.Length", "Petal.Width"))
  )
)

}
