% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{knit_print.reactable}
\alias{knit_print.reactable}
\title{Print a reactable widget for knitr}
\usage{
knit_print.reactable(x, ...)
}
\arguments{
\item{x}{A \code{\link[=reactable]{reactable()}} instance.}

\item{...}{Additional arguments passed to the S3 method.}
}
\description{
This S3 method exists to enable \code{\link[=reactable]{reactable()}}'s \code{static} rendering option.
}
\keyword{internal}
