% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\docType{package}
\name{reactable-package}
\alias{reactable-package}
\alias{_PACKAGE}
\title{reactable: Interactive Data Tables for R}
\description{
Interactive data tables for R, based on the 'React Table' JavaScript library. Provides an HTML widget that can be used in 'R Markdown' or 'Quarto' documents, 'Shiny' applications, or viewed from an R console.
}
\details{
See the \href{https://glin.github.io/reactable/}{online documentation} for
examples and an extensive usage guide.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://glin.github.io/reactable/}
  \item \url{https://github.com/glin/reactable}
  \item Report bugs at \url{https://github.com/glin/reactable/issues}
}

}
\author{
\strong{Maintainer}: Greg Lin \email{glin@glin.io}

Other contributors:
\itemize{
  \item Tanner Linsley (React Table library) [contributor, copyright holder]
  \item  Emotion team and other contributors (Emotion library) [contributor, copyright holder]
  \item Kent Russell (reactR package) [contributor, copyright holder]
  \item Ramnath Vaidyanathan (htmlwidgets package) [contributor, copyright holder]
  \item Joe Cheng (htmlwidgets package) [contributor, copyright holder]
  \item JJ Allaire (htmlwidgets package) [contributor, copyright holder]
  \item Yihui Xie (htmlwidgets package) [contributor, copyright holder]
  \item Kenton Russell (htmlwidgets package) [contributor, copyright holder]
  \item  Facebook, Inc. and its affiliates (React library) [contributor, copyright holder]
  \item  FormatJS (FormatJS libraries) [contributor, copyright holder]
  \item  Feross Aboukhadijeh, and other contributors (buffer library) [contributor, copyright holder]
  \item Roman Shtylman (process library) [contributor, copyright holder]
  \item James Halliday (stream-browserify library) [contributor, copyright holder]
  \item  Posit Software, PBC [funder, copyright holder]
}

}
\keyword{internal}
