% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{reactable}
\alias{reactable}
\title{Create a data table}
\usage{
reactable(data, rownames = FALSE, colnames = NULL, groupBy = NULL,
  columns = NULL, columnGroups = NULL, sortable = TRUE,
  resizable = FALSE, filterable = FALSE, defaultColDef = NULL,
  defaultSortOrder = "asc", defaultSorted = NULL, pagination = TRUE,
  defaultPageSize = 10, pageSizeOptions = c(10, 25, 50, 100),
  paginationType = "numbers", showPagination = NULL,
  showPageSizeOptions = TRUE, showPageInfo = TRUE, minRows = 1,
  selection = NULL, selectionId = NULL, details = NULL,
  outlined = FALSE, bordered = FALSE, borderless = FALSE,
  striped = FALSE, highlight = TRUE, compact = FALSE,
  showSortable = FALSE, class = NULL, style = NULL, inline = FALSE,
  width = "auto", height = "auto", elementId = NULL)
}
\arguments{
\item{data}{A data frame or matrix.}

\item{rownames}{Show row names?}

\item{colnames}{Optional named list of column names.}

\item{groupBy}{Optional vector of column names to group by.}

\item{columns}{Optional named list of column definitions. See \code{colDef()}.}

\item{columnGroups}{Optional list of column group definitions. See \code{colGroup()}.}

\item{sortable}{Enable sorting? Defaults to \code{TRUE}.}

\item{resizable}{Enable column resizing?}

\item{filterable}{Enable column filtering?}

\item{defaultColDef}{Default column definition used by every column. See \code{colDef()}.}

\item{defaultSortOrder}{Default sort order. Either \code{"asc"} for ascending
order or \code{"desc"} for descending order. Defaults to \code{"asc"}.}

\item{defaultSorted}{Optional vector of column names to sort by default.
Or to customize sort order, a named list with values of \code{"asc"} or \code{"desc"}.}

\item{pagination}{Enable pagination? Defaults to \code{TRUE}.}

\item{defaultPageSize}{Default page size for the table. Defaults to 10.}

\item{pageSizeOptions}{Page size options for the table. Defaults to 10, 25, 50, 100.}

\item{paginationType}{Pagination control to use. Either \code{"numbers"} for page
number buttons (the default), \code{"jump"} for a page jump, or \code{"simple"} to show
'Previous' and 'Next' buttons only.}

\item{showPagination}{Show pagination? Defaults to \code{TRUE} if the table has more
than one page.}

\item{showPageSizeOptions}{Show page size options?}

\item{showPageInfo}{Show page info?}

\item{minRows}{Minimum number of rows to show. Defaults to 1.}

\item{selection}{Enable row selection? Either \code{"multiple"} or \code{"single"} for
multiple or single row selection.}

\item{selectionId}{Shiny input ID for the selected rows. The selected rows are
represented as a vector of row indices, or \code{NULL} if no rows are selected.}

\item{details}{Additional content to display when expanding a row. A row details
definition or content renderer. See \code{rowDetails()}.}

\item{outlined}{Add borders around the table?}

\item{bordered}{Add borders around the table and every cell?}

\item{borderless}{Remove inner borders from table?}

\item{striped}{Add zebra-striping to table rows?}

\item{highlight}{Highlight table rows on hover? Defaults to \code{TRUE}.}

\item{compact}{Make tables more compact?}

\item{showSortable}{Show an indicator on sortable columns?}

\item{class}{Additional CSS classes to apply to the table.}

\item{style}{Inline styles to apply to the table. A named list or character string.}

\item{inline}{Display the table as an inline element, which shrinks to fit
its contents? By default, the table is displayed as a block element, which
expands to fit its parent container.}

\item{width}{Width in pixels (optional, defaults to automatic sizing).}

\item{height}{Height in pixels (optional, defaults to automatic sizing).}

\item{elementId}{Optional element ID for the widget.}
}
\value{
An htmlwidget.
}
\description{
Create a data table
}
