% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{resolvedData}
\alias{resolvedData}
\title{The result from handling a server-side data request}
\usage{
resolvedData(data, rowCount = NULL, maxRowCount = NULL)
}
\arguments{
\item{data}{The current page of data. A data frame.}

\item{rowCount}{The row count of the current page.}

\item{maxRowCount}{The maximum row count. Optional. Used to determine whether
the pagination bar should be kept visible when filtering or searching
reduces the current rows to one page, or when expanding rows
(when paginateSubRows is \code{TRUE}) would expand the table beyond one page.}
}
\description{
The result from handling a server-side data request
}
\seealso{
\code{\link[=reactableServerInit]{reactableServerInit()}} and \code{\link[=reactableServerData]{reactableServerData()}} for creating custom
server-side data backends.
}
