% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{rowDetails}
\alias{rowDetails}
\title{Row details definitions}
\usage{
rowDetails(render, html = FALSE, name = NULL, width = NULL)
}
\arguments{
\item{render}{Content renderer. A function that takes a row index argument
or a \code{JS()} function that takes a row info object as an argument.}

\item{html}{Render content as HTML? HTML strings are escaped by default.}

\item{name}{Expander column name.}

\item{width}{Expander column width in pixels.}
}
\description{
Row details definitions
}
