% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toJSON}
\alias{toJSON}
\title{Serialize JSON}
\usage{
toJSON(
  x,
  digits = getOption("reactable.json.digits", NA),
  func = getOption("reactable.json.func")
)
}
\arguments{
\item{digits}{Max number of digits to use for numeric values. Defaults to the \code{reactable.json.digits} option, or otherwise the maximum number of digits in jsonlite. Internal, undocumented, and unused. Only present because jsonlite's digits handling has changed between versions (1.8.5, then reverted in 1.8.7).}

\item{func}{Custom JSON serialization function. Experimental and for advanced use only. reactable may change how data is serialized between versions and does not guarantee stability of this feature.}
}
\description{
toJSON overrides the htmlwidgets default JSON serialization options for data:
}
\details{
\itemize{
\item Serialize numbers with max precision (typically 15 digits depending on the installed jsonlite version).
\item Preserve numeric NA, NaN, Inf, and -Inf as strings. String NAs are still serialized as \code{null}.
\item Serialize both datetimes and dates as ISO 8601 in UTC timezone.
}
}
\keyword{internal}
