% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{widget_html.reactable}
\alias{widget_html.reactable}
\title{Called by HTMLWidgets to produce the widget's root element}
\usage{
widget_html.reactable(id, style, class, ...)
}
\arguments{
\item{id}{Element ID.}

\item{style}{Element style.}

\item{class}{Element class.}

\item{...}{Additional arguments.}
}
\description{
Called by HTMLWidgets to produce the widget's root element
}
\keyword{internal}
