% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{colDef}
\alias{colDef}
\title{Column definitions}
\usage{
colDef(name = NULL, aggregate = NULL, sortable = NULL,
  resizable = NULL, filterable = NULL, show = TRUE,
  defaultSortOrder = NULL, format = NULL, cell = NULL,
  aggregated = NULL, footer = NULL, html = FALSE, showNA = FALSE,
  minWidth = NULL, maxWidth = NULL, width = NULL, align = NULL,
  class = NULL, style = NULL, headerClass = NULL,
  headerStyle = NULL, footerClass = NULL, footerStyle = NULL)
}
\arguments{
\item{name}{Column name.}

\item{aggregate}{Aggregate function. The name of a built-in aggregate
function or a custom \code{JS()} aggregate function. Built-in aggregate functions
are: \code{"mean"}, \code{"sum"}, \code{"max"}, \code{"min"}, \code{"count"}, \code{"unique"}, \code{"frequency"}.}

\item{sortable}{Enable sorting? Overrides the table option.}

\item{resizable}{Enable column resizing? Overrides the table option.}

\item{filterable}{Enable column filtering? Overrides the table option.}

\item{show}{Show the column? Defaults to \code{TRUE}.}

\item{defaultSortOrder}{Default sort order. Either \code{"asc"} for ascending
order or \code{"desc"} for descending order. Overrides the table option.}

\item{format}{Column formatting options. A \code{colFormat()} object to
format all cells, or a named list of \code{colFormat()} objects to format standard
cells (\code{"cell"}) and aggregated cells (\code{"aggregated"}) separately.}

\item{cell}{Custom cell renderer. A function that takes the cell value and row
index as arguments, or a \code{JS()} function that takes a cell info object as an
argument.}

\item{aggregated}{Custom aggregated cell renderer. A \code{JS()} function that takes
a cell info object as an argument.}

\item{footer}{Footer content or render function. Render functions can be an
R function that takes two arguments, the column values and column name, or a
\code{JS()} function that takes one argument, a column info object.}

\item{html}{Render cells as HTML? HTML strings are escaped by default.}

\item{showNA}{Show NA values? If \code{FALSE}, NA values will be left as empty cells.
Defaults to \code{FALSE}.}

\item{minWidth}{Min width of the column in pixels.}

\item{maxWidth}{Max width of the column in pixels.}

\item{width}{Fixed width of the column in pixels. Overrides minWidth and maxWidth.}

\item{align}{Column alignment. One of \code{"left"}, \code{"right"}, \code{"center"}.}

\item{class}{Additional CSS classes to apply to cells.}

\item{style}{Inline styles to apply to cells. A named list or character string.}

\item{headerClass}{Additional CSS classes to apply to the header.}

\item{headerStyle}{Inline styles to apply to the header. A named list or
character string.}

\item{footerClass}{Additional CSS classes to apply to the footer.}

\item{footerStyle}{Inline styles to apply to the footer. A named list or
character string.}
}
\description{
Column definitions
}
