% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{colDef}
\alias{colDef}
\title{Column definitions}
\usage{
colDef(name = NULL, aggregate = NULL, aggregated = NULL,
  sortable = NULL, resizable = NULL, filterable = NULL,
  show = TRUE, class = NULL, style = NULL, headerClass = NULL,
  headerStyle = NULL)
}
\arguments{
\item{name}{Column name.}

\item{aggregate}{Aggregate function name or JS callback.}

\item{aggregated}{Render function for aggregated cells.}

\item{sortable}{Enable sorting? Overrides the table option.}

\item{resizable}{Enable column resizing? Overrides the table option.}

\item{filterable}{Enable column filtering? Overrides the table option.}

\item{show}{Show the column? Defaults to \code{TRUE}.}

\item{class}{Additional CSS classes to apply to cells.}

\item{style}{Named list of inline styles to apply to cells.}

\item{headerClass}{Additional CSS classes to apply to the header.}

\item{headerStyle}{Named list of inline styles to apply to the header.}
}
\description{
Column definitions
}
