% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{colDef}
\alias{colDef}
\title{Column definitions}
\usage{
colDef(name = NULL, aggregate = NULL, aggregated = NULL,
  sortable = NULL, resizable = NULL, filterable = NULL,
  show = TRUE)
}
\arguments{
\item{name}{Column name}

\item{aggregate}{Aggregate function name or JS callback.}

\item{aggregated}{Render function for aggregated cells.}

\item{sortable}{Enable sorting? Overrides the table option.}

\item{resizable}{Enable column resizing? Overrides the table option.}

\item{filterable}{Enable column filtering? Overrides the table option.}

\item{show}{Show the column? Defaults to \code{TRUE}.}
}
\description{
Column definitions
}
