% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{colDef}
\alias{colDef}
\title{Column definitions}
\usage{
colDef(name = NULL, aggregate = NULL, sortable = NULL,
  resizable = NULL, filterable = NULL, show = TRUE,
  defaultSortOrder = NULL, render = NULL, renderAggregated = NULL,
  class = NULL, style = NULL, headerClass = NULL,
  headerStyle = NULL)
}
\arguments{
\item{name}{Column name.}

\item{aggregate}{Aggregate function name or JS callback.}

\item{sortable}{Enable sorting? Overrides the table option.}

\item{resizable}{Enable column resizing? Overrides the table option.}

\item{filterable}{Enable column filtering? Overrides the table option.}

\item{show}{Show the column? Defaults to \code{TRUE}.}

\item{defaultSortOrder}{Default sort order. Either \code{"asc"} for ascending
order or \code{"desc"} for descending order. Overrides the table option.}

\item{render}{Render function for standard cells.}

\item{renderAggregated}{Render function for aggregated cells.}

\item{class}{Additional CSS classes to apply to cells.}

\item{style}{Named list of inline styles to apply to cells.}

\item{headerClass}{Additional CSS classes to apply to the header.}

\item{headerStyle}{Named list of inline styles to apply to the header.}
}
\description{
Column definitions
}
