% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{colGroup}
\alias{colGroup}
\title{Column group definitions}
\usage{
colGroup(name, columns, headerClass = NULL, headerStyle = NULL)
}
\arguments{
\item{name}{Name of the column group.}

\item{columns}{Character vector of column names to group.}

\item{headerClass}{Additional CSS classes to apply to the header.}

\item{headerStyle}{Named list of inline styles to apply to the header.}
}
\description{
Column group definitions
}
