% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\docType{package}
\name{reactable-package}
\alias{reactable-package}
\alias{_PACKAGE}
\title{reactable: Interactive Data Tables Based on 'React Table'}
\description{
Interactive data tables for R, based on the 'React Table' JavaScript library. Provides an HTML widget that can be used in 'R Markdown' documents and 'Shiny' applications, or viewed from an R console.
}
\details{
See the \href{https://glin.github.io/reactable/}{online documentation} for
examples and an extensive usage guide.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://glin.github.io/reactable/}
  \item \url{https://github.com/glin/reactable}
  \item Report bugs at \url{https://github.com/glin/reactable/issues}
}

}
\author{
\strong{Maintainer}: Greg Lin \email{glin@glin.io}

Other contributors:
\itemize{
  \item Tanner Linsley (React Table library) [contributor, copyright holder]
  \item  Emotion team and other contributors (Emotion library) [contributor, copyright holder]
}

}
\keyword{internal}
