% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{reactable-server}
\alias{reactable-server}
\alias{reactableServerInit}
\alias{reactableServerData}
\title{Create custom server-side data backends for Shiny}
\usage{
reactableServerInit(
  x,
  data = NULL,
  columns = NULL,
  pageIndex = 0,
  pageSize = 0,
  sortBy = NULL,
  filters = NULL,
  searchValue = NULL,
  searchMethod = NULL,
  groupBy = NULL,
  pagination = NULL,
  paginateSubRows = NULL,
  selectedRowIds = NULL,
  expanded = NULL,
  ...
)

reactableServerData(
  x,
  data = NULL,
  columns = NULL,
  pageIndex = 0,
  pageSize = 0,
  sortBy = NULL,
  filters = NULL,
  searchValue = NULL,
  searchMethod = NULL,
  groupBy = NULL,
  pagination = NULL,
  paginateSubRows = NULL,
  selectedRowIds = NULL,
  expanded = NULL,
  ...
)
}
\arguments{
\item{x}{The server backend.}

\item{data}{The original table data. A data frame.}

\item{columns}{Table columns. A list of \code{\link[=colDef]{colDef()}} objects.}

\item{pageIndex}{The current page index. Starts at zero.}

\item{pageSize}{The current page size.}

\item{sortBy}{The current sorted columns. \code{NULL} if empty.}

\item{filters}{The current column filters. \code{NULL} if empty.}

\item{searchValue}{The current global search value. \code{NULL} if empty.}

\item{searchMethod}{The custom search method. A \code{\link[=JS]{JS()}} function.}

\item{groupBy}{The current grouped columns. \code{NULL} if empty.}

\item{pagination}{Whether pagination is enabled, \code{TRUE} or \code{FALSE}.}

\item{paginateSubRows}{Whether sub rows are paginated, \code{TRUE} or \code{FALSE}.}

\item{selectedRowIds}{The current selected rows.}

\item{expanded}{The current expanded rows.}

\item{...}{Additional arguments passed to the S3 method.}
}
\value{
\itemize{
\item \code{reactableServerData()} should return a \code{\link[=resolvedData]{resolvedData()}} object.
\item \code{reactableServerData()} should not return any value.
}
}
\description{
Custom server-side data backends are created using the
\href{https://adv-r.hadley.nz/s3.html}{S3 object system}.

To create a custom server-side data backend, provide an S3 object to the
\code{server} argument in \code{\link[=reactable]{reactable()}} with the following S3 methods defined:
\itemize{
\item \code{reactableServerInit} initializes the server backend (optional).
\item \code{reactableServerData} handles requests for data and should return a
\code{\link[=resolvedData]{resolvedData()}} object.
}

Custom backend methods do not have to accept every argument, and can choose
not to implement certain features such as grouping, row expansion, or
row selection.

If there is no server-side implementation for row expansion and row selection,
reactable will fall back to client-side row expansion and selection. This means
row expansion and selection will only work for rows on the current page, so for
example, selecting all rows in the table will only select rows on the current page.

Custom backend methods should accept additional arguments via \code{...} in case
new arguments are added in the future.
}
