% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{reactable}
\alias{reactable}
\title{Create a data table}
\usage{
reactable(data, rownames = FALSE, colnames = NULL, groupBy = NULL,
  columns = NULL, columnGroups = NULL, sortable = TRUE,
  resizable = TRUE, filterable = FALSE, defaultSortOrder = "asc",
  defaultSorted = NULL, defaultPageSize = 10, pageSizeOptions = c(10,
  25, 50, 100), showPagination = NULL, minRows = 1,
  selectable = FALSE, selectionType = "multiple", selectionId = NULL,
  outlined = FALSE, bordered = TRUE, striped = FALSE,
  highlight = TRUE, class = NULL, style = NULL, width = "auto",
  height = "auto", elementId = NULL)
}
\arguments{
\item{data}{A data frame or matrix.}

\item{rownames}{Show row names? Defaults to \code{FALSE}.}

\item{colnames}{Optional named list of column names.}

\item{groupBy}{Optional vector of column names to group by.}

\item{columns}{Optional named list of column definitions. See \code{colDef()}.}

\item{columnGroups}{Optional list of column group definitions. See \code{colGroup()}.}

\item{sortable}{Enable sorting? Defaults to \code{TRUE}.}

\item{resizable}{Enable column resizing? Defaults to \code{TRUE}.}

\item{filterable}{Enable column filtering? Defaults to \code{FALSE}.}

\item{defaultSortOrder}{Default sort order. Either \code{"asc"} for ascending
order or \code{"desc"} for descending order. Defaults to \code{"asc"}.}

\item{defaultSorted}{Optional vector of column names to sort by default.
Or to customize sort order, a named list with values of \code{"asc"} or \code{"desc"}.}

\item{defaultPageSize}{Default page size for the table. Defaults to 10.}

\item{pageSizeOptions}{Page size options for the table. Defaults to 10, 25, 50, 100.}

\item{showPagination}{Show pagination? Defaults to \code{TRUE} if the table fits on one page.}

\item{minRows}{Minimum number of rows to show. Defaults to 1.}

\item{selectable}{Enable row selection? Defaults to \code{FALSE}.}

\item{selectionType}{Row selection type. One of \code{"multiple"} (the default) or \code{"single"}.}

\item{selectionId}{Shiny input ID for the row selection. The row selection is
returned as a vector of row indices, or \code{NULL} if no rows are selected.}

\item{outlined}{Add an outline around the table? Defaults to \code{FALSE}.}

\item{bordered}{Add horizontal borders between table rows? Defaults to \code{TRUE}.}

\item{striped}{Add zebra-striping to table rows? Defaults to \code{FALSE}.}

\item{highlight}{Highlight table rows on hover? Defaults to \code{TRUE}.}

\item{class}{Additional CSS classes to apply to the table.}

\item{style}{Named list of inline styles to apply to the table.}

\item{width}{Width in pixels (optional, defaults to automatic sizing).}

\item{height}{Height in pixels (optional, defaults to automatic sizing).}

\item{elementId}{Optional element ID for the widget.}
}
\value{
An htmlwidget.
}
\description{
Create a data table
}
